Back

+++ src/main/java/com/example/security/ooredoo/repositories/InterventionRepo.java
@@ -3,7 +3,7 @@ package com.example.security.ooredoo.repositories;
 import com.example.security.ooredoo.entities.Intervention;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InterventionRepo extends JpaRepository<Intervention,Long> {
+public interface InterventionRepo extends JpaRepository<Intervention,Integer> {
 
 
 }
